backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
sqlalchemy>=2.0
psycopg[binary]>=3,<4
pydantic>=2.0
alembic>=1.13
python-dotenv>=1.0
